services/transceiver/src/requirements.txt
# Transceiver Python Requirements.

redis
discord
python-dotenv
